Track search event for all integrations
