+++ packages/client/components/ParabolScopingSearchBar.tsx
@@ -40,6 +40,7 @@ const ParabolScopingSearchBar = (props: Props) => {
         queryString={queryString ?? ''}
         meetingId={meetingId}
         linkedRecordName={'parabolSearchQuery'}
+        service={'parabol'}
       />
       <ParabolScopingSearchFilterToggle meeting={meeting} />
     </ScopingSearchBar>
